test: move files to modules
