+++ test/modules/airline.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
-import { Aircraft, faker } from '../src';
-import { seededTests } from './support/seededRuns';
-import { times } from './support/times';
+import { Aircraft, faker } from '../../src';
+import { seededTests } from './../support/seededRuns';
+import { times } from './../support/times';
 
 const NON_SEEDED_BASED_RUN = 5;
