+++ test/modules/finance_iban.spec.ts
@@ -1,8 +1,8 @@
 import validator from 'validator';
 import { describe, expect, it } from 'vitest';
-import { faker } from '../src';
-import ibanLib from '../src/modules/finance/iban';
-import { times } from './support/times';
+import { faker } from '../../src';
+import ibanLib from '../../src/modules/finance/iban';
+import { times } from './../support/times';
 
 const NON_SEEDED_BASED_RUN = 25;
