+++ test/modules/number.spec.ts
@@ -1,7 +1,7 @@
 import validator from 'validator';
 import { describe, expect, it } from 'vitest';
-import { faker, FakerError } from '../src';
-import { seededTests } from './support/seededRuns';
+import { faker, FakerError } from '../../src';
+import { seededTests } from './../support/seededRuns';
 
 describe('number', () => {
   seededTests(faker, 'number', (t) => {
