+++ test/modules/system.spec.ts
@@ -1,8 +1,8 @@
 import validator from 'validator';
 import { describe, expect, it } from 'vitest';
-import { faker, fakerSK } from '../src';
-import { seededTests } from './support/seededRuns';
-import { times } from './support/times';
+import { faker, fakerSK } from '../../src';
+import { seededTests } from './../support/seededRuns';
+import { times } from './../support/times';
 
 const NON_SEEDED_BASED_RUN = 5;
